+++ driven/storage/model_auth.go
@@ -1,6 +1,9 @@
 package storage
 
-import "time"
+import (
+	"core-building-block/core/model"
+	"time"
+)
 
 type loginSession struct {
 	ID string `bson:"_id"`
@@ -36,4 +39,20 @@ type loginSession struct {
 
 	DateUpdated *time.Time `bson:"date_updated"`
 	DateCreated time.Time  `bson:"date_created"`
+}
+
+type serviceAccount struct {
+	AccountID string `bson:"account_id"`
+	Name      string `bson:"name"`
+
+	AppID *string `bson:"app_id"`
+	OrgID *string `bson:"org_id"`
+
+	Permissions []model.Permission `bson:"permissions"`
+	FirstParty  bool               `bson:"first_party"`
+
+	Credentials []model.ServiceAccountCredential `bson:"credentials"`
+
+	DateCreated time.Time  `bson:"date_created"`
+	DateUpdated *time.Time `bson:"date_updated"`
 }