+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Expose revoke permissions from role Admin API [#416](https://github.com/rokwire/core-building-block/issues/416)
+- Service accounts [#306](https://github.com/rokwire/core-building-block/issues/306)
 
 ## [1.22.0] - 2022-04-02
 ### Added